Add page reducer
