+++ packages/noya-state/src/reducers/pageReducer.ts
@@ -1,0 +1,121 @@
+import Sketch from '@sketch-hq/sketch-file-format-ts';
+import produce from 'immer';
+import { uuid } from 'noya-renderer';
+import * as Layers from '../layers';
+import * as Models from '../models';
+import { getCurrentPage, getCurrentPageIndex } from '../selectors/selectors';
+import { UUID } from '../types';
+import { ApplicationState } from './applicationReducer';
+
+export type PageAction =
+  | [type: 'movePage', sourceIndex: number, destinationIndex: number]
+  | [type: 'selectPage', pageId: UUID]
+  | [type: 'addPage', name: string]
+  | [type: 'deletePage']
+  | [type: 'renamePage', name: string]
+  | [type: 'duplicatePage'];
+
+export function pageReducer(
+  state: ApplicationState,
+  action: PageAction,
+): ApplicationState {
+  switch (action[0]) {
+    case 'selectPage': {
+      return produce(state, (draft) => {
+        draft.selectedPage = action[1];
+      });
+    }
+    case 'addPage': {
+      const [, name] = action;
+
+      return produce(state, (draft) => {
+        const pages = draft.sketch.pages;
+        const user = draft.sketch.user;
+
+        const newPage: Sketch.Page = produce(Models.page, (page) => {
+          page.do_objectID = uuid();
+          page.name = name || `Page ${pages.length + 1}`;
+          return page;
+        });
+
+        user[newPage.do_objectID] = {
+          scrollOrigin: '{0, 0}',
+          zoomValue: 1,
+        };
+
+        pages.push(newPage);
+
+        draft.selectedPage = newPage.do_objectID;
+      });
+    }
+    case 'renamePage': {
+      const [, name] = action;
+      const pageIndex = getCurrentPageIndex(state);
+
+      return produce(state, (draft) => {
+        const pages = draft.sketch.pages;
+        const page = pages[pageIndex];
+
+        pages[pageIndex] = produce(page, (page) => {
+          page.name = name || `Page ${pages.length + 1}`;
+          return page;
+        });
+      });
+    }
+    case 'duplicatePage': {
+      const pageIndex = getCurrentPageIndex(state);
+
+      return produce(state, (draft) => {
+        const pages = draft.sketch.pages;
+        const user = draft.sketch.user;
+        const page = pages[pageIndex];
+
+        const duplicatePage = produce(page, (page) => {
+          page.name = `${page.name} Copy`;
+
+          Layers.visit(page, (layer) => {
+            layer.do_objectID = uuid();
+            if (layer.style) layer.style.do_objectID = uuid();
+          });
+
+          return page;
+        });
+
+        user[duplicatePage.do_objectID] = {
+          scrollOrigin: user[page.do_objectID].scrollOrigin,
+          zoomValue: user[page.do_objectID].zoomValue,
+        };
+
+        pages.push(duplicatePage);
+        draft.selectedPage = duplicatePage.do_objectID;
+      });
+    }
+    case 'deletePage': {
+      const page = getCurrentPage(state);
+      const pageIndex = getCurrentPageIndex(state);
+
+      return produce(state, (draft) => {
+        const pages = draft.sketch.pages;
+        const user = draft.sketch.user;
+
+        delete user[page.do_objectID];
+        pages.splice(pageIndex, 1);
+
+        const newIndex = Math.max(pageIndex - 1, 0);
+        draft.selectedPage = pages[newIndex].do_objectID;
+      });
+    }
+    case 'movePage': {
+      const [, sourceIndex, destinationIndex] = action;
+
+      return produce(state, (draft) => {
+        const sourceItem = draft.sketch.pages[sourceIndex];
+
+        draft.sketch.pages.splice(sourceIndex, 1);
+        draft.sketch.pages.splice(destinationIndex, 0, sourceItem);
+      });
+    }
+    default:
+      return state;
+  }
+}
